+++ changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## Version 0.3.0 (2023/02/22)
+
+- Added verses manipulation
+- Added verses tests
+- Added examples in readme
+- More fixes
+
+---
+
 ## Version 0.2.0 (2023/02/21)
 
 - Moved tests to their own file